requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==4.0
django-environ==0.8.1
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.1
